requirements.txt
appnope==0.1.3
asn1crypto==1.5.1
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.12.2
boto3==1.26.115
botocore==1.29.115
certifi==2022.12.7
charset-normalizer==3.1.0
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
greenlet==2.0.2
idna==3.4
install==1.3.5
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
ipython-sql==0.5.0
jedi==0.18.2
jmespath==1.0.1
jupyter_client==8.2.0
jupyter_core==5.3.0
lxml==4.9.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.2
packaging==23.1
pandas==2.0.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.2.0
prettytable==3.7.0
prompt-toolkit==3.0.38
psutil==5.9.4
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.0
pyodbc==4.0.39
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pyzmq==25.0.2
redshift-connector==2.0.910
requests==2.28.2
s3transfer==0.6.0
scramp==1.4.4
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.9
sqlparse==0.4.3
stack-data==0.6.2
tornado==6.2
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wcwidth==0.2.6
